server/requirements.txt
defusedxml==0.4.1
Django==1.8.5
django-allauth==0.23.0
django-cors-headers==1.1.0
django-encrypted-fields==1.1.2
django-extensions==1.7.5
django-picklefield==0.3.2
django-rest-auth==0.5.0
djangorestframework==3.2.4
oauthlib==1.0.3
Pillow==3.4.2
pyasn1==0.1.9
pycrypto==2.6.1
python-keyczar==0.716
python-openid==2.2.5
python3-openid==3.0.10
requests==2.8.0
requests-oauthlib==0.5.0
six==1.10.0
